Add: Script to automate some of the pre-run setup
The new `run-jetmon.sh` script is now the command run by `Dockerfile`.
It ensures that the directories are present and generates the key and
cert if they are not present.

Changes were made to the code in the repo to apply the changes as
described in the testing instructions as they are necessary to run the
codebase currently.