Testing: Exporatory code. Do not merge.
Testing out workers sending their details to the parent to allow the
parent to more intelligently distribute work while minimizing the number
of idle workers. Setting the foundation for a smarter way to recycle
workers that is more reliable and doesn't lose pending URL checks on the
recycled worker. Added a port 7802 listener that dumps stats about Jetmon
and its workers.