Enhancement: Aggregate StatsD metrics
Per a request from Systems, the worker check stats are now aggregated
and sent to the parent process. This increases efficiency by not
requiring each worker to communicate with the StatsD lib and
dramatically reduces the number of StatsD metrics that are sent.

The rtt values are now tracked as gauge metrics that are aggregated to
just the avg, max, and min values for each stats interval.