Fix: Inflated SPS metric calculations
The SPS calculation did not account for the stats flush interval set by
the STATS_UPDATE_INTERVAL_MS config. This meant that for a configured
flush every 5 seconds, that the reported SPS value was inflated 5x. This
change fixes the calculation.

While making this change, I also noticed that a site success count was
incorrectly reported as a total count. This metric is now changed to a
success count. A new total count metric is added to provide a correct
total count.